PE/lab03/src/requirements.txt
annotated-types
anyio
argon2-cffi
argon2-cffi-bindings
Brotli
certifi
cffi
click
fastapi
greenlet
h11
h2
hpack
hyperframe
idna
minio
nats-py==2.9.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser
pycryptodome
pydantic
pydantic_core
PyJWT
PySocks
python-multipart
PyYAML
redis
setuptools==75.8.0
sniffio
SQLAlchemy==2.0.40
starlette
typing_extensions
urllib3
uvicorn
wheel==0.45.1
zstandard==0.23.0
